esercizio

+++ Back-End/MESE5/Settimana1/Giorno2/esercizio/src/main/java/NextDevs/esercizio/bean/Tavolo.java
@@ -14,5 +14,5 @@ public class Tavolo {
     private int numeroCopertiMax;
     private StatoTavolo statoTavolo;
     @Value("${costo.coperto}")
-    private int costoCoperto;
+    private Double costoCoperto;
 }
